Modified Producer and Consumer test
